Extract the editor wrapper into separate component
